fix: Unable to link secondary auth provider on custom domain
