+++ app/scenes/Settings/Authentication.tsx
@@ -19,6 +19,7 @@ import useRequest from "~/hooks/useRequest";
 import useStores from "~/hooks/useStores";
 import SettingRow from "./components/SettingRow";
 import { setPostLoginPath } from "~/hooks/useLastVisitedPath";
+import { getRedirectUrl } from "~/utils/urls";
 import { settingsPath } from "~/utils/routeHelpers";
 import DomainManagement from "./components/DomainManagement";
 import Button from "~/components/Button";
@@ -97,7 +98,7 @@ function Authentication() {
 
   const handleConnectProvider = React.useCallback((name: string) => {
     setPostLoginPath(settingsPath("authentication"));
-    window.location.href = `/auth/${name}?host=${window.location.host}`;
+    window.location.href = getRedirectUrl(`/auth/${name}`);
   }, []);
 
   const handleToggleGroupSync = React.useCallback(
